backend/db/requirements-sync.txt
supabase==2.5.0
pandas==2.2.2
numpy==1.26.4
pyarrow==16.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sshtunnel==0.4.0
paramiko==3.4.0
